+++ pkg/go/graph/weighted_graph.go
@@ -1,0 +1,94 @@
+package graph
+
+import (
+	"fmt"
+
+	openfgav1 "github.com/openfga/api/proto/openfga/v1"
+	"gonum.org/v1/gonum/graph/encoding"
+	"gonum.org/v1/gonum/graph/encoding/dot"
+	"gonum.org/v1/gonum/graph/multi"
+)
+
+type WeightedAuthorizationModelGraph struct {
+	*multi.DirectedGraph
+	drawingDirection DrawingDirection
+}
+
+func NewWeightedAuthorizationModelGraph(model *openfgav1.AuthorizationModel) (*WeightedAuthorizationModelGraph, error) {
+	g, err := NewAuthorizationModelGraph(model)
+	if err != nil {
+		return nil, err
+	}
+	g, err = g.Reversed() // we want direction of Check
+	if err != nil {
+		return nil, err
+	}
+
+	graphBuilder := &WeightedAuthorizationModelGraphBuilder{multi.NewDirectedGraph()}
+
+	iterEdges := g.Edges()
+	for iterEdges.Next() {
+		nextEdge, ok := iterEdges.Edge().(multi.Edge)
+		if !ok {
+			return nil, fmt.Errorf("%w: could not cast %v to multi.Edge", ErrBuildingGraph, iterEdges.Edge())
+		}
+
+		// NOTE: because we use a multigraph, one edge can include multiple lines, so we need to add each line individually.
+		iterLines := nextEdge.Lines
+		for iterLines.Next() {
+			nextLine := iterLines.Line()
+			castedEdge, ok := nextLine.(*AuthorizationModelEdge)
+			if !ok {
+				return nil, fmt.Errorf("%w: could not cast %v to AuthorizationModelEdge", ErrBuildingGraph, nextLine)
+			}
+
+			err = graphBuilder.AddEdgeWithWeights(castedEdge)
+			if err != nil {
+				return nil, err
+			}
+		}
+	}
+
+	err = graphBuilder.AssignWeights()
+	if err != nil {
+		return nil, err
+	}
+
+	multigraph, ok := graphBuilder.DirectedMultigraphBuilder.(*multi.DirectedGraph)
+	if !ok {
+		return nil, fmt.Errorf("%w: could not cast to DirectedGraph", ErrBuildingGraph)
+	}
+
+	wg := &WeightedAuthorizationModelGraph{multigraph, g.drawingDirection}
+
+	return wg, nil
+}
+
+var _ dot.Attributers = (*WeightedAuthorizationModelGraph)(nil)
+
+func (wb *WeightedAuthorizationModelGraph) DOTAttributers() (encoding.Attributer, encoding.Attributer, encoding.Attributer) {
+	return wb, nil, nil
+}
+
+func (wb *WeightedAuthorizationModelGraph) Attributes() []encoding.Attribute {
+	rankdir := "BT" // bottom to top
+	if wb.drawingDirection == DrawingDirectionCheck {
+		rankdir = "TB" // top to bottom
+	}
+
+	return []encoding.Attribute{{
+		Key:   "rankdir", // https://graphviz.org/docs/attrs/rankdir/
+		Value: rankdir,
+	}}
+}
+
+// GetDOT returns the DOT visualization. The output text is stable.
+// It should only be used for debugging.
+func (wb *WeightedAuthorizationModelGraph) GetDOT() string {
+	dotRepresentation, err := dot.MarshalMulti(wb, "", "", "")
+	if err != nil {
+		return ""
+	}
+
+	return string(dotRepresentation)
+}
